refactor: optimize token selector sorting with improved data structure
- Move sort types from kit-bg to shared/types for better architecture
- Change assetCtxs from array to Record<number, IPerpsAssetCtx> for O(1) lookup
- Add case-insensitive name sorting with sensitivity: 'base'
- Remove align prop and simplify SortableHeaderCell component
- Remove sortPerpsAssetIndices from default export
- Use nullish coalescing operator for better type safety
