+++ packages/kit/src/states/jotai/contexts/hyperliquid/atoms.ts
@@ -43,9 +43,9 @@ export const {
 
 export const { atom: perpsAllAssetCtxsAtom, use: usePerpsAllAssetCtxsAtom } =
   contextAtom<{
-    assetCtxs: HL.IPerpsAssetCtx[];
+    assetCtxs: Record<number, HL.IPerpsAssetCtx>;
   }>({
-    assetCtxs: [],
+    assetCtxs: {},
   });
 
 export const { atom: l2BookAtom, use: useL2BookAtom } =
